test(1396): default JWT_SECRET for all unit-test trees that drive the OAuth callback (fix pre-push regression)
